+++ src/HmIPAccessory.ts
@@ -17,11 +17,11 @@ export class HmIPAccessory {
   public register() {
     if (this.isFromCache) {
       this.log.debug('Updating accessory: %s (%s) -> uuid %s',
-        this.accessory.context.device.label, this.accessory.context.device.id, this.accessory.UUID);
+        this.accessory.displayName, this.accessory.context.device.id, this.accessory.UUID);
       this.api.updatePlatformAccessories([this.accessory]);
     } else {
       this.log.info('Register accessory: %s (%s) -> uuid %s',
-        this.accessory.context.device.label, this.accessory.context.device.id, this.accessory.UUID);
+        this.accessory.displayName, this.accessory.context.device.id, this.accessory.UUID);
       this.api.registerPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [this.accessory]);
     }
   }