+++ src/devices/HmIPGenericDevice.ts
@@ -3,14 +3,39 @@ import {CharacteristicGetCallback, PlatformAccessory, Service} from 'homebridge'
 import {HmIPPlatform} from '../HmIPPlatform';
 import {HmIPDevice, HmIPGroup} from '../HmIPState';
 
+interface SupportedOptionalFeatures {
+  IFeatureDeviceParticulateMatterSensorCommunicationError: boolean;
+  IFeatureDeviceCoProRestart: boolean;
+  IFeatureDeviceOverheated: boolean;
+  IOptionalFeatureDutyCycle: boolean;
+  IFeatureMulticastRouter: boolean;
+  IFeatureDeviceCoProUpdate: boolean;
+  IFeaturePowerShortCircuit: boolean;
+  IFeatureDevicePowerFailure: boolean;
+  IFeatureDeviceTemperatureHumiditySensorCommunicationError: boolean;
+  IFeatureShortCircuitDataLine: boolean;
+  IFeatureRssiValue: boolean;
+  IFeatureBusConfigMismatch: boolean;
+  IFeatureDeviceUndervoltage: boolean;
+  IFeatureDeviceParticulateMatterSensorError: boolean;
+  IFeatureDeviceOverloaded: boolean;
+  IFeatureDeviceCoProError: boolean;
+  IFeatureDeviceIdentify: boolean;
+  IOptionalFeatureLowBat: boolean;
+  IOptionalFeatureMountingOrientation: boolean;
+  IFeatureDeviceTemperatureHumiditySensorError: boolean;
+  IFeatureDeviceTemperatureOutOfRange: boolean;
+}
+
 interface DeviceBaseChannel {
   functionalChannelType: string;
   unreach: boolean;
   lowBat: boolean;
   rssiDeviceValue: number;
   rssiPeerValue: number;
-  dutyCycle: false;
-  configPending: false;
+  dutyCycle: boolean;
+  configPending: boolean;
+  supportedOptionalFeatures: SupportedOptionalFeatures;
 }
 
 /**
@@ -20,11 +45,11 @@ export abstract class HmIPGenericDevice {
 
   protected unreach = false;
   protected lowBat = false;
-  protected rssiDeviceValue = 0.0;
+  protected rssiDeviceValue = 0;
   protected rssiPeerValue = 0;
   protected dutyCycle = false;
   protected configPending = false;
-  private batteryService: Service;
+  private readonly batteryService: Service | undefined;
 
   protected constructor(
     protected readonly platform: HmIPPlatform,
@@ -36,25 +61,29 @@ export abstract class HmIPGenericDevice {
       .setCharacteristic(this.platform.Characteristic.SerialNumber, accessory.context.device.id)
       .setCharacteristic(this.platform.Characteristic.FirmwareRevision, accessory.context.device.firmwareVersion);
 
-    this.batteryService = this.accessory.getService(this.platform.Service.BatteryService) || this.accessory.addService(this.platform.Service.BatteryService)!;
-    this.batteryService.getCharacteristic(this.platform.Characteristic.BatteryLevel) // this is actually optional since iOS 14
-      .on('get', this.handleBatteryLevelGet.bind(this));
-    this.batteryService.getCharacteristic(this.platform.Characteristic.ChargingState) // this is actually optional since iOS 14
-      .on('get', this.handleChargingStateGet.bind(this));
-    this.batteryService.getCharacteristic(this.platform.Characteristic.StatusLowBattery)
-      .on('get', this.handleStatusLowBatteryGet.bind(this));
-  }
+    const hmIPDevice = <HmIPDevice>accessory.context.device;
+    let featureLowBat = false;
 
-  handleBatteryLevelGet(callback: CharacteristicGetCallback) {
-    callback(null, (this.lowBat ? 0 : 100));
-  }
+    for (const id in hmIPDevice.functionalChannels) {
+      const channel = hmIPDevice.functionalChannels[id];
+      if (channel.functionalChannelType === 'DEVICE_OPERATIONLOCK' || channel.functionalChannelType === 'DEVICE_BASE') {
+        const baseChannel = <DeviceBaseChannel>channel;
 
-  handleChargingStateGet(callback: CharacteristicGetCallback) {
-    callback(null, this.platform.Characteristic.ChargingState.NOT_CHARGEABLE);
+        featureLowBat = baseChannel.supportedOptionalFeatures.IOptionalFeatureLowBat;
+        break;
+      }
+    }
+
+    if (featureLowBat) {
+      this.batteryService = this.accessory.getService(this.platform.Service.Battery) || this.accessory.addService(this.platform.Service.Battery)!;
+      this.batteryService.getCharacteristic(this.platform.Characteristic.StatusLowBattery)
+        .on('get', this.handleStatusLowBatteryGet.bind(this));
+    }
   }
 
   handleStatusLowBatteryGet(callback: CharacteristicGetCallback) {
-    callback(null, (this.lowBat ? this.platform.Characteristic.StatusLowBattery.BATTERY_LEVEL_LOW : this.platform.Characteristic.StatusLowBattery.BATTERY_LEVEL_NORMAL));
+    callback(null, (this.lowBat ? this.platform.Characteristic.StatusLowBattery.BATTERY_LEVEL_LOW
+      : this.platform.Characteristic.StatusLowBattery.BATTERY_LEVEL_NORMAL));
   }
 
   protected updateDevice(hmIPDevice: HmIPDevice, groups: { [key: string]: HmIPGroup }) {
@@ -64,17 +93,16 @@ export abstract class HmIPGenericDevice {
         const baseChannel = <DeviceBaseChannel>channel;
 
         if (baseChannel.unreach != null && baseChannel.unreach !== this.unreach) {
-          this.platform.log.info(`Unreach of ${this.accessory.displayName} changed to ${baseChannel.unreach}`);
           this.unreach = baseChannel.unreach;
+          this.platform.log.info('Unreach of %s changed to %s', this.accessory.displayName, this.unreach);
         }
 
-        if (baseChannel.lowBat !== null && baseChannel.lowBat !== this.lowBat) {
-          this.platform.log.info(`LowBat of ${this.accessory.displayName} changed to ${baseChannel.lowBat}`);
+        if (this.batteryService && baseChannel.lowBat !== null && baseChannel.lowBat !== this.lowBat) {
           this.lowBat = baseChannel.lowBat;
+          this.platform.log.info('LowBat of %s changed to %s', this.accessory.displayName, this.lowBat);
           this.batteryService.setCharacteristic(this.platform.Characteristic.StatusLowBattery,
             this.lowBat ? this.platform.Characteristic.StatusLowBattery.BATTERY_LEVEL_LOW
               : this.platform.Characteristic.StatusLowBattery.BATTERY_LEVEL_NORMAL);
-          this.batteryService.setCharacteristic(this.platform.Characteristic.BatteryLevel, this.lowBat ? 0 : 100);
         }
       }
     }